+++ pywr/_core.pyx
@@ -218,32 +218,24 @@ cdef class ScenarioIndex:
 
 
 cdef class Timestep:
-    def __init__(self, datetime, int index, double days):
-        self._datetime = pd.Timestamp(datetime)
-        self._index = index
-        self._days = days
-        tt = self.datetime.timetuple()
-        self.dayofyear = tt.tm_yday
-        self.day = tt.tm_mday
-        self.month = tt.tm_mon
-        self.year = tt.tm_year
+    def __init__(self, period, int index, double days):
+        self.period = period
+        self.index = index
+        if days <= 0:
+            raise ValueError("The days argument must be > 0.")
+        self.days = days
+        self.dayofyear = period.dayofyear
+        self.day = period.day
+        self.month = period.month
+        self.year = period.year
 
     property datetime:
         """Timestep representation as a `datetime.datetime` object"""
         def __get__(self, ):
-            return self._datetime
-
-    property index:
-        """The index of the timestep for use in arrays"""
-        def __get__(self, ):
-            return self._index
-
-    property days:
-        def __get__(self, ):
-            return self._days
+            return self.period.to_timestamp()
 
     def __repr__(self):
-        return "<Timestep date=\"{}\">".format(self._datetime.strftime("%Y-%m-%d"))
+        return "<Timestep date=\"{}\">".format(self.period.strftime("%Y-%m-%d"))
 
 cdef class Domain:
     """ Domain class which all Node objects must have. """
@@ -966,7 +958,7 @@ cdef class Storage(AbstractStorage):
         cdef ScenarioIndex si
 
         for i, si in enumerate(self.model.scenarios.combinations):
-            self._volume[i] += self._flow[i]*ts._days
+            self._volume[i] += self._flow[i]*ts.days
             # Ensure variable maximum volume is taken in to account
 
             mxv = self.get_max_volume(si)
@@ -1031,7 +1023,7 @@ cdef class AggregatedStorage(AbstractStorage):
             for s in self._storage_nodes:
                 self._flow[i] += s._flow[i]
                 mxv += s.get_max_volume(si)
-            self._volume[i] += self._flow[i]*ts._days
+            self._volume[i] += self._flow[i]*ts.days
 
             # Ensure variable maximum volume is taken in to account
             try: