+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All issue numbers are relative to https://github.com/pywr/pywr/issues unless otherwise stated.
 
+
+## v1.3.0
+
+### New Features
+
+- Allow use of parameters as values in `ControlCurveInterpolatedParameter` (#750)
+- Added `ScenarioWrapper` parameter (#763)
+
+### Bug fixes
+
+- Fix loading PiecewiseLink with parameters from JSON (#749)
+- Fixed a bug with `CSVRecorder` not saving volumes correctly (#767)
+
+### Miscellaneous
+
+- Removed `six` as dependency (#745)
+- Removed `pywr.__git_hash__` (#752)
+- Removed `Blender` node (#757)
+
 ## v1.2.0
 
 ### New Features